counterfeit_requirements.txt
opencv-python==4.11.0.86
numpy==2.3.0
torch==2.7.1+cu128
torchvision==0.22.1+cu128
transformers==4.52.4
Pillow==11.0.0
requests==2.32.4
scikit-learn==1.6.2
scikit-image==0.24.0
matplotlib==3.10.3
python-dotenv==1.1.0
dataclasses-json==0.6.7 